Add request details test
This adds a new page that when requested, will output the details of the
request. It was created as part of working on [Add ability to set
user-agent to use](https://github.com/DEFRA/quke/pull/60) in **Quke**.

Along with it is a feature that provides an example of how you can then
test a page for specific content. Admittedly not an amazing addition as
an example, but it gives meaning for the page being there, and is a
working example nonetheless.