Fix javascript error tests (#5)
The javascript error tests only worked if you were using Phantomjs for the browser (and therefore the poltergiest Driver).

Quke has removed support for Phantomjs so we can no longer rely on it for an example of testing for JavaScript errors.

This change updates the example to use one that relies on Chrome being used as the driver. Again support is not universal or consistent.