Fix request tests
The request tests assume that the user agent will be specified in the config file, and then check we get a matching value.

However for this to work anyone running this project needed to remember to add the expected value to the `.config.yml` they created locally.

In hindsight the easiest way to resolve the issue and make things simpler going forward was to add a correctly setup `.config.yml` file to the project.