Update project to refer to saved .config.yml
Previously the project expected the user to create the `.config.yml` file. Now we have one committed to the project for convenience, setup to ensure the tests should pass.

So this change updates the README content to reflect this, and adds in the notes and reasoning for some settings to `.config.yml`.