Remove google based tests
The intention of quke-example is to demonstrate how to use Quke in an environment that is stable and   does not rely on external web sites.

For awhile now we've had a feature that was named as a way of testing support for Browserstack, but    was actually just performing a Google search for `browserstack`.

This was a standin, but as we don't really have a way to test browserstack integration other than      running the features through browserstack using real credentials there is no point in maintaining this
standin.

So this change removes the feature, and brings the project inline with its original intent.