Add support for custom config files (#8)
Now we have added `.config.yml` to source control, any changes made to it will be picked up by git. To support running the tests with different config eg changing the driver from Chrome to Firefox, it'd be cleaner to do it in a separate file.

So this change adds an entry to gitignore to support this, and a brief explanation to the README that explains how to do it.