Prepare for move to Defra
As the defacto example for the Defra [Quke](https://github.com/DEFRA/quke) framework and with working CI implemented, we are now moving this out of its holding place against a personal account and into the Defra organisation.

Main change is the LICENSE, and the bottom section of the README to match in with our normal standards.