Switch to using gem for the demo app
To make this example project as close to what it would actually be like as possible, we created [quke-demo-app](https://github.com/DEFRA/quke-demo-app) which is a gem that hosts the demo site.

This means we can remove the whole `demo_app` folder and all special references to it, with just a simple line in the Gemfile.