Fix rubocop issues
Not sure what was happening before, but spotted that Travis was failing because it could not find rubocop. Turns out this is as expected because we don't list rubocop in our gemfile.

So this change sorts out the travis config so we can still run rubocop and lint the project, plus fixes any issues found.